+++ interview/filters.py
@@ -1,7 +1,7 @@
 import django_filters
 
 from interview.models import Process
-from ref.models import Subsidiary
+from ref.models import Subsidiary, Consultant
 
 
 class ProcessFilter(django_filters.FilterSet):
@@ -18,3 +18,9 @@ class ProcessSummaryFilter(django_filters.FilterSet):
 class InterviewSummaryFilter(django_filters.FilterSet):
     subsidiary = django_filters.ModelChoiceFilter(queryset=Subsidiary.objects.all(), field_name="process__subsidiary")
     last_state_change = django_filters.DateFromToRangeFilter(field_name="planned_date")
+
+
+class InterviewListFilter(django_filters.FilterSet):
+    subsidiary = django_filters.ModelChoiceFilter(queryset=Subsidiary.objects.all(), field_name="process__subsidiary")
+    last_state_change = django_filters.DateFromToRangeFilter(field_name="planned_date")
+    interviewer = django_filters.ModelChoiceFilter(queryset=Consultant.objects.all(), field_name="interviewers")