+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # v1.14.0
 - Update django to 3.2.11
+- New page: list past and future interviews
 
 # v1.13.0 (2022-01-26)
 - Add interview.prequalification to interviews TSV export